html/semantics/forms/the-input-element/input-setcustomvalidity.html

<!DOCTYPE HTML>
<title>input setCustomValidity</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<input id='test'>

<script>

test(() => {
  let input = document.getElementById("test");
  assert_false(input.validity.customError);
  input.setCustomValidity("custom error");
  assert_true(input.validity.customError);
}, "input setCustomValidity is correct")

</script>
